error handling

+++ backend/routes/posts.js
@@ -36,8 +36,10 @@ router.post("" , checkAuth ,multer({storage : storage}).single("image") ,(req, r
   const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
+   
     post.save().then(createdPost => {
       res.status(201).json({
         message: "Post added successfully",
@@ -48,6 +50,11 @@ router.post("" , checkAuth ,multer({storage : storage}).single("image") ,(req, r
           imagePath: createdPost.imagePath
         }
       });
+    }).catch(err => {
+      res.status(500).json({
+        message: "Creating post failed!"
+      }
+      )
     });
     
   });
@@ -65,12 +72,22 @@ router.post("" , checkAuth ,multer({storage : storage}).single("image") ,(req, r
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
       });
       console.log(post);
-      Post.updateOne({ _id: req.params.id }, post).then(result => {
-        res.status(200).json({ message: "Update successful!" });
-      });
+      Post.updateOne({ _id: req.params.id , creator: req.userData.userId }, post).then(result => {
+        if(result.nModified > 0){
+          res.status(200).json({ message: "Update successful!" });
+        } else {
+          res.status(401).json({ message: "Not authorised!" });
+        }
+       
+      }).catch(err => {
+        res.status(500).json({
+          message: "Couldn't update post failed!"
+        })
+      });;
     }
   );
   
@@ -94,6 +111,11 @@ router.post("" , checkAuth ,multer({storage : storage}).single("image") ,(req, r
         posts: fetchedPosts,
         maxPosts: count
       });
+    }).catch(err => {
+      res.status(500).json({
+        message: "Fetching post failed"
+      }
+      )
     });
   });
   
@@ -105,16 +127,28 @@ router.post("" , checkAuth ,multer({storage : storage}).single("image") ,(req, r
       }else{
         res.status(404).json({message: 'Post not found!!!'})
       }
-    })
+    }).catch(err => {
+      res.status(500).json({
+        message: "Fetching post failed"
+      })
+    });
   })
   
   
   
   router.delete("/:id" , checkAuth  , (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-      console.log(result);
+    Post.deleteOne({ _id: req.params.id , creator: req.userData.userId }).then(result => {
+     if(result.n > 0){
       res.status(200).json({ message: "Post deleted!" });
-    });
+     } else {
+      res.status(401).json({ message: "Not authorised!" });
+     }
+      
+    }).catch(err => {
+      res.status(500).json({
+        message: "Deleting post failed"
+      })
+    });;
   });
